feat: Merge branch dev.
libdata: Major updates to nc* functions following cf conventions.

functools: new mustBe* validation functions, fixed dealout, struct2varargin, logical flag feature in parseoptarg, rmpairs option in parseparampairs.

libarrays: nancat feature, refactored ensembleList accepts struct input and returns allcombos.

libhydro: hydraulic geometry function, fix / clean up readMerra2.

libplot: Refactor defaultcolors to wrap distinguishable_colors, add defaultlinestyles function, make color optional input for fillplot and fix errors in Props settings.

libraster: Refine rasterref and rasterize and warn they produce different results for missing pixels. Add gridxyz function. Fix ncrowcol, plotraster.

libspatial: Major updates to updateCoordinates, loadprojcrs function, interpolationPoints and resampleCoordinates reconciled with runoff.

+++ libsys/backupfile.m
@@ -44,8 +44,13 @@ function [fullpath_bk, filename_bk] = backupfile(filename, makecopy, makezip)
    fullpath_bk = fullfile(filepath, filename_bk);
 
    if makecopy
+      if ~isfile(fullpath) && ~isfolder(fullpath)
+         % This lets backupfile be called without if isfile() in the caller
+         warning('No backup made. File not found: %s', fullpath)
+         return
+      end
       if isfile(fullpath_bk) || isfolder(fullpath_bk)
-         warning('Backup already exists. No copy made.');
+         warning('No backup made. Backup already exists: %s', fullpath_bk);
       else
          try
             copyfile(fullpath, fullpath_bk);
@@ -66,13 +71,6 @@ function [fullpath_bk, filename_bk] = backupfile(filename, makecopy, makezip)
          fprintf('Backup created: %s\n', fullpath_bk);
       end
    end
-end
-
-function filedate = mkfiledate(dateformat)
-   if nargin < 1
-      dateformat = 'dd-MMM-yyyy_HH-mm-ss';
-   end
-   filedate = strrep(char(datetime('now', 'Format', dateformat)), '-', '');
 end
 
 function filename = rmtrailingsep(filename)
